requirements.txt
flask
flask-cors
requests
pywebview
duckduckgo_search
